[BREAKING] Add ember-try for TS 3.6, 3.8, 3.9, drops 3.5
This matches the upcoming support matrix in https://github.com/machty/ember-concurrency/pull/357

This drops TS 3.5 as a tested configuration. This is inevitable,
since we will start depending on the types from upstream instead
of vendoring our own, and the upstream types does not support 3.5.

in comparasion even more breaking than this change, so it shouldn't
be an issue to drop 3.5 here.