Fix IE11 compatibility (avoid Object.entries).
This library is used by @ember/test-waiters's test suite, and
`@ember/test-waiters` **must** function on IE11 throughout the Ember 3.x
timeframe. `Object.entries` is not present on IE11, causing failures in
both `@ember/test-waiters` and `@ember/test-helpers` when running under
IE11.

Thankfully, migrating to `Object.keys` + `for..of` is pretty straight
forward and with this change the test suite passes on IE11.